CODE/requirements.txt
comtypes==1.1.11
MouseInfo==0.1.3
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
pyttsx3==2.90
pytweening==1.0.4
pywin32==304
thread6==0.2.0
tk==0.1.0
